adding time_integral variable including get and set function and fixing some types

+++ main/include/eudaq/WaveformSignalRegion.hh
@@ -27,8 +27,8 @@ class WaveformSignalRegion:public TObject {
         void SetPolarity(signed char pol){polarity=pol;}
         void SetPulserPolarity(signed char pol){pulserPolarity=pol;}
         void AddIntegral(WaveformIntegral integralDef);
-        uint32_t GetLowBoarder() {return low_border;}
-        uint32_t GetHighBoarder() {return high_border;}
+        uint16_t GetLowBoarder() {return low_border;}
+        uint16_t GetHighBoarder() {return high_border;}
         uint16_t GetPeakPosition(){return peak_pos;}
         void SetName(std::string name){this->name = name;}
         const char *GetName(){return name.c_str();}
@@ -38,8 +38,8 @@ class WaveformSignalRegion:public TObject {
         std::string name;
         signed char polarity;
         signed char pulserPolarity;
-        uint32_t low_border;
-        uint32_t high_border;
+        uint16_t low_border;
+        uint16_t high_border;
         uint16_t peak_pos;
         std::map<std::string,WaveformIntegral> integrals;
         std::vector<std::string> integral_names;
